+++ testing/testing_ztrsm_vbatched.cpp
@@ -405,7 +405,7 @@ int main( int argc, char** argv)
                 }
                 
                 printf("  %-10lld  %-5lld %-5lld  %-7.2f ( %-7.2f )  %-7.2f ( %-7.2f )   %-8.2e        %-8.2e  %s\n",
-                        batchCount, max_M, max_N,
+                        (long long)batchCount, (long long)max_M, (long long)max_N,
                         magma_perf,  1000.*magma_time,
                         cpu_perf,    1000.*cpu_time,
                         magma_error, lapack_error, 
@@ -413,7 +413,7 @@ int main( int argc, char** argv)
             }
             else {
                 printf("  %-10lld %-5lld %-5lld  %-7.2f ( %-7.2f )    ---   (   ---   )      ---           ---  \n",
-                        batchCount, max_M, max_N,
+                        (long long)batchCount, (long long)max_M, (long long)max_N,
                         magma_perf,  1000.*magma_time);
             }
             magma_free_cpu( h_A );
